Fix C4005 '_WIN32_WINNT' macro redefinition warning for cotire unity builds:
VS2015's comes with Windows SDK 8.1 so it defaults _WIN32_WINNT to target Windows 8.1.  We want to target Windows 7.1, so we need to override the default target by explicitly setting our desired target via command-line options.  Trying to do so in platform.h is too late for cotire unity builds.  For non-cotire unity builds, doing so in platform.h is okay, but for cotire-unity builds we need to do it via the command-line (due to the way it works with respect to constructing a pre-compiled header to be used for the building the rest of the project).  NOTE: This change is only in order to be consistent across all External Packages since decNumber doesn't actually use cotire unity builds.  It uses cotire yes (for pre-compiled headers), but not cotire UNITY builds since they don't work for some reason and I haven't figured out why yet.  But as I said, I want to be consistent.